requirements.txt
langchain-core
langchain-community
langchain_groq
faiss-cpu
langchain_google_genai 
langchain_anthropic 
langchain_huggingface
langchain_openai
pypdf
dotenv
bs4
ipython
pprintpp
